Struct rust_hypervisor_firmware_subset::mem::MemoryRegion

pub struct MemoryRegion { /* private fields */ }

Provides a checked way to access memory offsets from a range of raw memory

Implementations

impl MemoryRegion

pub const fn new(base: u64, length: u64) -> MemoryRegion

pub fn from_bytes(data: &mut [u8]) -> MemoryRegion

Take a slice and turn it into a region of memory

pub fn as_bytes(&mut self) -> &mut [u8]

pub fn as_mut_slice<T>(&mut self, offset: u64, length: u64) -> &mut [T]

Expose a section of the memory region as a slice

pub fn read_u8(&self, offset: u64) -> u8

Read a single byte at a given offset

pub fn read_u16(&self, offset: u64) -> u16

Read a single word at a given offset

pub fn read_u32(&self, offset: u64) -> u32

Read a single dword at a given offset

pub fn read_u64(&self, offset: u64) -> u64

pub fn write<T>(&self, offset: u64, value: T)

Write a value at the given offset

pub fn write_u8(&self, offset: u64, value: u8)

Write a single byte at given offset

pub fn write_u16(&self, offset: u64, value: u16)

Write a single word at given offset

pub fn write_u32(&self, offset: u64, value: u32)

Write a single dword at given offset

pub fn write_u64(&self, offset: u64, value: u64)

Write a single qword at given offset

pub fn io_read_u8(&self, offset: u64) -> u8

Read a single byte at given offset with a mechanism suitable for MMIO

pub fn io_read_u16(&self, offset: u64) -> u16

Read a single word at given offset with a mechanism suitable for MMIO

pub fn io_read_u32(&self, offset: u64) -> u32

Read a single dword at given offset with a mechanism suitable for MMIO

pub fn io_read_u64(&self, offset: u64) -> u64

Read a single qword at given offset with a mechanism suitable for MMIO

pub fn io_write_u8(&self, offset: u64, value: u8)

Write a single byte at given offset with a mechanism suitable for MMIO

pub fn io_write_u16(&self, offset: u64, value: u16)

Write a single word at given offset with a mechanism suitable for MMIO

pub fn io_write_u32(&self, offset: u64, value: u32)

Write a single dword at given offset with a mechanism suitable for MMIO

pub fn io_write_u64(&self, offset: u64, value: u64)

Write a single qword at given offset with a mechanism suitable for MMIO